Created a Gradle Project named VisArtApplication
New Changes:
The project has been configured with the proper Spring Boot and Hibernate
dependencies installed so far. Basic JUnit testing added as dependency
by default as well. gitignore has been filled in as well. The project was
configured in Eclipse, so it can be imported directly into that easily.

Next Step:
Test this project with Hibernate and Spring Boot commands to ensure that
the dependencies successfully install (after gradlew build of course)